+++ src/app.module.ts
@@ -4,6 +4,8 @@ import { AppService } from './app.service';
 import { SystemModule } from './system/system.module';
 import { BrowserModule } from './browser/browser.module';
 
+import { TypeOrmModule } from '@nestjs/typeorm';
+
 import configuration from './configuration/configuration';
 
 import { ConfigModule } from '@nestjs/config';
@@ -14,7 +16,12 @@ import { ConfigurationModule } from './configuration/configuration.module';
       isGlobal: true,
       load: [configuration],
       envFilePath: ['.env']
-  }), SystemModule, BrowserModule, ConfigurationModule],
+  }), SystemModule, BrowserModule, ConfigurationModule, TypeOrmModule.forRoot({
+    type: 'sqlite',
+    database: __dirname + '/../database/database.sqlite',
+    entities: [__dirname + '/**/*.entity{.ts,.js}'],
+    synchronize: true,
+  })],
   controllers: [AppController],
   providers: [AppService],
 })