+++ src/main.ts
@@ -2,11 +2,23 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ConfigService } from '@nestjs/config';
 
+import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
+
+const packageJson = require('../package.json');
+
 declare const module: any;
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   const configService = app.get(ConfigService);
+
+  const options = new DocumentBuilder()
+    .setTitle('PRINT API')
+    .setDescription(packageJson.description)
+    .setVersion(packageJson.version)
+    .build();
+  const document = SwaggerModule.createDocument(app, options);
+  SwaggerModule.setup('api', app, document);
   
   await app.listen(configService.get('port'));