+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-078-05
+### Changed
+- Instead of reading the filesystem the checks will run against a sqlite database
+
+### Added
+- Support for automatic delete after request
+- Counting of downloads
+- Support for sqlite
+- Swagger Documentation (/api)
+
 ## [0.2.3] - 2020-07-30
 ### Added
 - File entry will also return created_at and updated_at